refactor: 重构 skills 目录结构，解耦合同自动编号模块
- 创建合同处理工作流目录结构
- 将 contract_numbering.py 拆分为多个模块：
  - formats.py: 格式定义
  - detector.py: 编号检测
  - converter.py: 编号转换
  - utils.py: 工具函数
  - cli.py: 命令行入口
  - __init__.py: 入口导出
  - __main__.py: 模块执行入口
- 新增 CLAUDE.md 开发规范
- 更新 README.md 和 CHANGELOG.md 文档
