+++ skills/合同处理/CHANGELOG.md
@@ -1,0 +1,30 @@
+# Changelog - 合同处理工作流
+
+所有对此工作流的更改都将记录在此文件中。
+
+格式基于 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.0.0/)，
+
+## [1.1.0] - 2026-07-31
+
+### 新增
+
+- 创建合同处理工作流目录结构
+- 合同自动编号 skill 升级到 v1.1.0
+  - 支持纯数字格式：1. 1.1 1.1.1 1.1.1.1 1.1.1.1.1
+  - 用户选择功能：调用时询问用户选择编号格式
+  - 代码重构：拆分为多个模块
+
+### 工作流变更
+
+- 创建 README.md 工作流说明
+- 创建 CHANGELOG.md 变更日志
+
+## [1.0.0] - 2026-07-31
+
+### 新增
+
+- 合同自动编号 skill v1.0.0
+  - 支持一、1.（1）① 四级编号格式
+  - 智能识别文档结构
+  - 智能推断缺失编号
+  - 跳过签字盖章部分
