+++ skills/合同处理/合同自动编号/CHANGELOG.md
@@ -14,7 +14,12 @@
 
 ### 变更
 
-- 重构代码结构，支持多种编号格式
+- 重构代码结构，拆分为多个模块：
+  - `formats.py`：格式定义
+  - `detector.py`：编号检测
+  - `converter.py`：编号转换
+  - `utils.py`：工具函数
+  - `cli.py`：命令行入口
 - 更新 README 文档，说明两种格式的使用方法
 
 ## [1.0.0] - 2026-07-31
