+++ skills/合同处理/合同自动编号/__main__.py
@@ -1,0 +1,6 @@
+"""Allow running the skill as a module."""
+
+from .cli import main
+
+if __name__ == '__main__':
+    main()
